core: Add split_logger/SplitDrain and ElasticLoggingConfig struct
