+++ core/src/lib.rs
@@ -19,5 +19,6 @@ mod log;
 mod subgraph;
 
 pub use graphql::GraphQlRunner;
-pub use log::{elastic_logger, ElasticDrainConfig};
+pub use log::elastic::{elastic_logger, ElasticDrainConfig, ElasticLoggingConfig};
+pub use log::split::split_logger;
 pub use subgraph::{SubgraphInstanceManager, SubgraphProvider, SubgraphProviderWithNames};
